Modified dma gen
